start adding --input-simulation-treefname
and rearrange treegenerator

+++ python/recombinator.py
@@ -63,7 +63,9 @@ class Recombinator(object):
                 self.mute_models[region][model][parameter_name] = line['value']
         treegen = treegenerator.TreeGenerator(args, self.shm_parameter_dir, seed=seed)
         self.treefname = self.workdir + '/trees.tre'
-        treegen.generate_trees(seed, self.treefname)  # NOTE not really a newick file, since I hack on the per-region branch length info at the end of each line
+        if self.args.input_simulation_treefname is not None:
+            utils.simplerun('cp %s %s' % (self.args.input_simulation_treefname, self.treefname), debug=False)
+        treegen.generate_trees(seed, self.treefname, self.workdir)  # NOTE not really a newick file, since I hack on the per-region branch length info at the end of each line
         with open(self.treefname, 'r') as treefile:  # read in the trees (and other info) that we just generated
             self.treeinfo = treefile.readlines()
         os.remove(self.treefname)
@@ -72,7 +74,10 @@ class Recombinator(object):
 
     # ----------------------------------------------------------------------------------------
     def __del__(self):
-        os.rmdir(self.workdir)
+        if len(os.listdir(self.workdir)) == 0:
+            os.rmdir(self.workdir)
+        else:
+            print '  couldn\'t exit cleanly, workdir %s not empty' % self.workdir
 
     # ----------------------------------------------------------------------------------------
     def read_insertion_content(self):
